+++ SerialPrograms/Source/PokemonSV/PokemonSV_Panels.cpp
@@ -35,7 +35,7 @@
 #include "Programs/FastCodeEntry/PokemonSV_ClipboardFastCodeEntry.h"
 #include "Programs/FastCodeEntry/PokemonSV_VideoFastCodeEntry.h"
 
-#include "Programs/ShinyHunting/PokemonSV_ShinyHunt-AreaZeroPlatform.h"
+#include "Programs/ShinyHunting/PokemonSV_ShinyHunt-AreaZyroPlatform.h"
 
 #include "Programs/Glitches/PokemonSV_RideCloner-1.0.1.h"
 #include "Programs/Glitches/PokemonSV_CloneItems-1.0.1.h"
@@ -98,7 +98,7 @@ std::vector<PanelEntry> PanelListFactory::make_panels() const{
 
     if (PreloadSettings::instance().DEVELOPER_MODE){
         ret.emplace_back("---- Shiny Hunting ----");
-        ret.emplace_back(make_single_switch_program<ShinyHuntAreaZeroPlatform_Descriptor, ShinyHuntAreaZeroPlatform>());
+        ret.emplace_back(make_single_switch_program<ShinyHuntAreaZyroPlatform_Descriptor, ShinyHuntAreaZyroPlatform>());
     }
 
     ret.emplace_back("---- Glitches ----");
